Preserve SteamID64 for server GC messages

+++ csgo_gc/steam_hook.cpp
@@ -1142,7 +1142,7 @@ static void Hk_SteamGameServer_RunCallbacks()
                 break;
 
             case HostEvent::NetMessage:
-                s_serverGC->m_networking.SendMessage((uint32_t)event.id, event.buffer.data(), static_cast<uint32_t>(event.buffer.size()));
+                s_serverGC->m_networking.SendMessage(event.id, event.buffer.data(), static_cast<uint32_t>(event.buffer.size()));
                 break;
 
             default:
